Complete redo of gloo & interfacehook
Interface Hook is gone & now merged into glop; this is because igloo
was importing too many pages previously, which severely slows down the
page load. In the future potentially don’t import any libraries until
it’s determined that the page that we’re on is WP:Gloo/run? Also,
cachebypass was added to external URLs, which should make life easier
(a lot easier) in the future